Split WorkspaceReducer into domain-scoped reducers
- Split the 563-line `WorkspaceReducer` into four domain reducers + a shared helpers namespace, keeping
  `WorkspaceReducer.reduce` as a thin dispatcher.
  - Collapsed duplicate `.selectProject` / `.selectWorktree` cases since their bodies were identical.
  - No behavior changes — all 29 `WorkspaceReducerTests` pass unchanged.
